% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{calculer_moyenne_buts}
\alias{calculer_moyenne_buts}
\title{Calculer la moyenne des buts par equipe}
\usage{
calculer_moyenne_buts(data)
}
\arguments{
\item{data}{Un tableau de donnees}
}
\value{
Un tableau resume
}
\description{
Calculer la moyenne des buts par equipe
}
