% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{filtrer_poste}
\alias{filtrer_poste}
\title{Filtrer les joueurs par poste}
\usage{
filtrer_poste(data, poste_choisi)
}
\arguments{
\item{data}{Un tableau de donnees}

\item{poste_choisi}{Le poste a filtrer (ex: "FW")}
}
\value{
Un tableau filtre
}
\description{
Filtrer les joueurs par poste
}
