% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{tracer_buts}
\alias{tracer_buts}
\title{Tracer le graphique des buts par equipe}
\usage{
tracer_buts(data_resume)
}
\arguments{
\item{data_resume}{Les donnees resumees}
}
\value{
Un graphique ggplot2
}
\description{
Tracer le graphique des buts par equipe
}
